+++ clang/term/ext.c
@@ -1,3 +1,3 @@
-fn u16 term_ext(Term t) {
+fn u32 term_ext(Term t) {
   return (t >> EXT_SHIFT) & EXT_MASK;
 }
