+++ clang/term/new.c
@@ -1,4 +1,4 @@
-fn Term term_new(u8 sub, u8 tag, u16 ext, u64 val) {
+fn Term term_new(u8 sub, u8 tag, u32 ext, u64 val) {
   return ((u64)sub << SUB_SHIFT)
        | ((u64)(tag & TAG_MASK) << TAG_SHIFT)
        | ((u64)(ext & EXT_MASK) << EXT_SHIFT)
