+++ clang/term/new/dp0.c
@@ -1,3 +1,3 @@
-fn Term term_new_dp0(u16 lab, u64 loc) {
+fn Term term_new_dp0(u32 lab, u64 loc) {
   return term_new(0, DP0, lab, loc);
 }
