+++ clang/term/new/dp1.c
@@ -1,3 +1,3 @@
-fn Term term_new_dp1(u16 lab, u64 loc) {
+fn Term term_new_dp1(u32 lab, u64 loc) {
   return term_new(0, DP1, lab, loc);
 }
