+++ clang/term/new/nam.c
@@ -1,3 +1,3 @@
-fn Term term_new_nam(u16 nam) {
+fn Term term_new_nam(u32 nam) {
   return term_new(0, NAM, nam, 0);
 }
