+++ clang/term/new/pri.c
@@ -1,7 +1,7 @@
-fn Term term_new_pri_at(u64 loc, u16 prim, u32 ari, Term *args) {
+fn Term term_new_pri_at(u64 loc, u32 prim, u32 ari, Term *args) {
   return term_new_at(loc, PRI, prim, ari, args);
 }
 
-fn Term term_new_pri(u16 prim, u32 ari, Term *args) {
+fn Term term_new_pri(u32 prim, u32 ari, Term *args) {
   return term_new_pri_at(heap_alloc(ari), prim, ari, args);
 }
