+++ clang/term/new/ref.c
@@ -1,3 +1,3 @@
-fn Term term_new_ref(u16 nam) {
+fn Term term_new_ref(u32 nam) {
   return term_new(0, REF, nam, 0);
 }
