+++ clang/term/new/swi.c
@@ -1,5 +1,5 @@
 // SWI: λ{num: f; g} - number switch (same as MAT but for parser/printer)
-fn Term term_new_swi(u16 num, Term f, Term g) {
+fn Term term_new_swi(u32 num, Term f, Term g) {
   u64 loc = heap_alloc(2);
   heap_set(loc + 0, f);
   heap_set(loc + 1, g);
