+++ clang/wnf/alo_nod.c
@@ -1,10 +1,10 @@
 // @{s} T{a,b,...}
 // ---------------- ALO-NOD
 // T{@{s}a, @{s}b, ...}
-fn Term wnf_alo_nod(u64 alo_loc, u64 ls_loc, u16 len, Term book) {
+fn Term wnf_alo_nod(u64 alo_loc, u64 ls_loc, u32 len, Term book) {
   u64 loc = term_val(book);
   u8  tag = term_tag(book);
-  u16 ext = term_ext(book);
+  u32 ext = term_ext(book);
   u32 ari = term_arity(book);
   Term args[16];
   if (ari == 0) {
